fix: StampsSQL 결과에 DTO 매핑

+++ GTA_SPRING/src/main/java/com/jsj/GTA/domain/stamps/UsersStampCountDto.java
@@ -5,9 +5,9 @@ import lombok.Getter;
 @Getter
 public class UsersStampCountDto {
     private Long id;
-    private int count;
+    private Long count;
 
-    public UsersStampCountDto(Long id, int count) {
+    public UsersStampCountDto(Long id, Long count) {
         this.id = id;
         this.count = count;
     }
